+++ lime_tbx/presentation/cli/cli.py
@@ -37,6 +37,7 @@ from lime_tbx.common.datatypes import (
     SatellitePoint,
     SurfacePoint,
     EocfiPath,
+    MultipleCustomPoint,
 )
 from lime_tbx.common import constants, logger
 from lime_tbx.common.constants import CompFields
@@ -49,6 +50,7 @@ from lime_tbx.application.simulation.comparison import comparison
 from lime_tbx.application.simulation.comparison.utils import (
     sort_by_mpa,
     average_comparisons,
+    filter_out_3sigmas_iter,
 )
 from lime_tbx.application.filedata import moon, srf as srflib, csv, lglod as lglodlib
 from lime_tbx.application.coefficients.update import Update
@@ -123,6 +125,8 @@ def print_help():
     - `-t, --timeseries`: Use a CSV file with multiple datetimes.
     - `-C, --coefficients`: Change the coefficients version used by the toolbox.
     - `-i, --interpolation-settings`: Modify interpolation settings via JSON input.
+    - `--filter3sigma`: Only applies for comparisons: Iteratively filters out data with a channel-wise
+        relative difference of 3 or more times the standard deviation from the mean.
 
     The function also provides examples of valid input formats and highlights
     any constraints or dependencies for specific options.
@@ -143,17 +147,22 @@ comparisons for some given observations files in GLOD format.\n"
     print("  -u, --update\t\t Updates the coefficients.")
     print("  -e, --earth\t\t Performs simulations from a geographic point.")
     print(f"\t\t\t -e lat_deg,lon_deg,height_km,datetime_isoformat")
-    print("  -l, --lunar\t\t Performs a simulation from a selenographic point.")
+    print("  -l, --lunar\t\t Performs a simulation from a selenographic point/s.")
     print(
         "\t\t\t -l distance_sun_moon,distance_observer_moon,selen_obs_lat,selen_obs_lon,\
 selen_sun_lon,moon_phase_angle"
     )
+    print("\t\t\t -l multiple_points.csv")
     print("  -s, --satellite\t Performs simulations from a satellite point.")
     print(f"\t\t\t -s sat_name,datetime_isoformat")
     print(
         "  -c, --comparison\t Performs comparisons from observations files in GLOD format."
     )
     print('\t\t\t -c "input_glod1.nc input_glod2.nc ..."')
+    print(
+        "\t\t\t Accepts the flag --filter3sigma: Iteratively filters out data with a channel-wise"
+        " relative difference of 3 or more times the standard deviation from the mean."
+    )
     print("  -o, --output\t\t Select the output path and format.")
     print("\t\t\t If it's a simulation:")
     print(f"\t\t\t   GRAPH: -o graph,{imsel},refl,irr,dolp,aolp")
@@ -182,8 +191,8 @@ selen_sun_lon,moon_phase_angle"
 in GLOD format."
     )
     print(
-        "  -t, --timeseries\t Select a file with multiple datetimes instead of \
-inputing directly only one datetime. Valid only if the main option is -e or -s."
+        "  -t, --timeseries\t Select a file with multiple datetimes instead of"
+        " inputing directly only one datetime. Valid only if the main option is -e or -s."
     )
     print(
         "  -C, --coefficients\t Change the coefficients version used by the TBX, \
@@ -707,6 +716,33 @@ class CLI:
         self._calculate_all(point)
         self.exporter.export(point, export_data, self.srf)
 
+    def calculate_multiselenographic(
+        self, input_path: str, export_data: export.ExportData
+    ):
+        """Runs the simulation for multiple selenographic coordinates
+
+        Parameters
+        ----------
+        input_path: str
+            Path of the csv file where the coordinates are specified
+        export_data : export.ExportData
+            The export configuration.
+
+        Raises
+        ------
+        ExportError
+            If something wrong happens during export.
+        """
+        try:
+            cps = csv.read_selenopoints(input_path)
+        except Exception as e:
+            raise ParsingError(
+                f"Error reading multiple selenographic input file: {str(e)}"
+            ) from e
+        point = MultipleCustomPoint(cps)
+        self._calculate_all(point)
+        self.exporter.export(point, export_data, self.srf)
+
     def _add_observation(self, obs: LunarObservation):
         for i, pob in enumerate(self.loaded_moons):
             if obs.dt < pob.dt:
@@ -718,6 +754,7 @@ class CLI:
         self,
         input_files: List[str],
         ed: export.ExportComparison,
+        filter3sigma: bool,
     ):
         """Performs comparisons between simulation results and observational data.
 
@@ -727,6 +764,8 @@ class CLI:
             List of file paths containing observational data.
         ed : export.ExportComparison
             The comparison export configuration.
+        filter3sigma: bool
+            Flag indicating if the a 3 sigma filter should be iteratively applied.
 
         Raises
         ------
@@ -769,6 +808,8 @@ class CLI:
         co = comparison.Comparison(self.kernels_path)
         cimel_coef = self.settings_manager.get_cimel_coef()
         comps = co.get_simulations(mos, self.srf, cimel_coef, self.lime_simulation)
+        if filter3sigma:
+            comps = filter_out_3sigmas_iter(comps)
         # EXPORT
         skip_uncs = self.settings_manager.is_skip_uncertainties()
         if isinstance(ed, export.ExportNetCDF):
@@ -1145,6 +1186,7 @@ Run 'lime -h' for help."
         """
         export_data: export.ExportData = None
         timeseries = None
+        filter3sigma = False
         # Check if it's comparison
         is_comparison = any(item[0] in ("-c", "--comparison") for item in opts)
         mod_interp_settings = False
@@ -1180,6 +1222,8 @@ Run 'lime -h' for help."
                 elif opt in ("-i", "--interpolation-settings"):
                     self._parse_interp_settings(arg)
                     mod_interp_settings = True
+                elif opt in ("--filter3sigma"):
+                    filter3sigma = True
         except CLIError as e:
             eprint(str(e))
             return 1
@@ -1244,11 +1288,15 @@ Run 'lime -h' for help."
                     break
                 elif opt in ("-l", "--lunar"):  # Lunar
                     params_str = arg.split(",")
-                    if len(params_str) != 6:
+                    if len(params_str) == 1:
+                        multiseleno_path = params_str[0]
+                        self.calculate_multiselenographic(multiseleno_path, export_data)
+                    elif len(params_str) == 6:
+                        params = list(map(float, params_str))
+                        self.calculate_selenographic(*params, export_data)
+                    else:
                         eprint("Error: Wrong number of arguments for -l")
                         return 1
-                    params = list(map(float, params_str))
-                    self.calculate_selenographic(*params, export_data)
                     break
                 elif opt in ("-c", "--comparison"):  # Comparison
                     params = args
@@ -1257,12 +1305,12 @@ Run 'lime -h' for help."
                     input_files = []
                     for param in params:
                         input_files += glob.glob(param)
-                    self.calculate_comparisons(input_files, export_data)
+                    self.calculate_comparisons(input_files, export_data, filter3sigma)
                     break
         except export.ExportError as e:
             eprint(str(e))
             return 1
-        except LimeException as e:
+        except (CLIError, LimeException) as e:
             eprint(f"Error: {str(e)}")
             return 1
         except Exception as e: