+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [//]: # "## [unreleased] - yyyy-mm-dd"
 
+## [1.3.0] - 2025-09-19
+
+### Added
+
+- Filter 3σ: Option in the Comparison page to filter out measurements that are more than 3σ away
+  from the mean relative difference. Filtering is done channel-wise.
+- Multiple Selenographic Points: Support for loading multiple selenographic coordinates at once
+  via CSV file input in the selenographic input tab.
+
+### Changed
+
+- Updated EO-CFI library dependencies to version 4.29 and recompiled EO-CFI dependent binaries.
+
 ## [1.2.2] - 2025-07-21
 
 ### Changed
@@ -82,7 +95,6 @@ correctly in such cases.
 - Timeseries input file now also accepts regular format timestamps, not only CSV.
 - Added METOP-A, METOP-B and METOP-C data (TLE/3LE). METOP-B and METOP-C will need to be periodically updated.
 - Updated TLE/3LE data for PLEIADES 1A, PLEIADES 1B and PROBA-V, now covering the period from their launch until 2025-04-14.
-- Updated EO-CFI library dependencies to version 4.28.
 
 ### Changed
 
@@ -90,6 +102,7 @@ correctly in such cases.
 - Provide support for Python 3.11 and 3.12.
 - Updated library dependencies and versions.
 - Refactored main Python package grouping subpackages in layer architecture based packages.
+- Updated EO-CFI library dependencies to version 4.28 and recompiled EO-CFI dependent binaries.
 
 ### Fixed
 
@@ -113,7 +126,9 @@ calculations compared to v1.0.3. (**NFR306**)
 
 Initial version that serves as the baseline for tracking changes in the change log.
 
-[unreleased]: https://github.com/LIME-ESA/lime_tbx/compare/v1.2.1...HEAD
+[unreleased]: https://github.com/LIME-ESA/lime_tbx/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/LIME-ESA/lime_tbx/compare/v1.2.2...1.3.0
+[1.2.2]: https://github.com/LIME-ESA/lime_tbx/compare/v1.2.1...1.2.2
 [1.2.1]: https://github.com/LIME-ESA/lime_tbx/compare/v1.2.0...1.2.1
 [1.2.0]: https://github.com/LIME-ESA/lime_tbx/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/LIME-ESA/lime_tbx/compare/v1.0.3...v1.1.0